% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayAlltraits.R
\name{displayAllTraits}
\alias{displayAllTraits}
\title{displayAllTraits}
\usage{
displayAllTraits(takes = NA)
}
\arguments{
\item{takes}{: Takes no parameter}
}
\value{
a dataframe whose one column in traits, another is their emoticon relator; can be saved in an object
}
\description{
Displays all of the traits and their unique emoticon associations. You must use the traits name to plot your emojis.
}
\examples{
displayAllTraits(takes= NA)

}
