% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmojiFromTheTrait.R
\name{getEmojiFromTheTrait}
\alias{getEmojiFromTheTrait}
\title{getEmojiFromTheTrait}
\usage{
getEmojiFromTheTrait(aTrait)
}
\arguments{
\item{aTrait}{: Takes a single trait}
}
\value{
Returns an utf-8 encoded emoji; can be saved in an object (Won't Render on Windows!!)
}
\description{
Returns a single emoji for ONE trait only.
}
\examples{
getEmojiFromTheTrait(aTrait= "Happy")

}
