% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyTraits.R
\name{plotMyTraits}
\alias{plotMyTraits}
\title{plotMyTraits}
\usage{
plotMyTraits(traits)
}
\arguments{
\item{traits}{: Accepts a character vector of length 4, which contains your unique traits}
}
\value{
If no error, you would get a banner size plot saved in your Desktop. File name is "mydna.png"
}
\description{
User can pass in 4 single traits of a character type inside a "c()" function & a banner will be generated & saved in the Desktop.
}
\examples{
plotMyTraits(c("happy","american","heaven","seeker"))

}
