% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitsLookup.R
\name{traitsLookup}
\alias{traitsLookup}
\title{traitsLookup}
\usage{
traitsLookup(takes = NA)
}
\arguments{
\item{takes}{: Takes no parameter}
}
\value{
A character vector containing traits; can be saved in an object
}
\description{
Takes user keywords via prompt & returns appropriate traits. Users get input prompt in the console where they put 4 keywords & function gives back traits, if possible.
}
\examples{
traitsLookup(takes= NA)

}
