% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMyTraits.R
\name{plotMyTraits}
\alias{plotMyTraits}
\title{Plot the traits in their emoji representation}
\usage{
plotMyTraits(traits)
}
\arguments{
\item{c("trait1", "trait2", "trait3", "trait4")}{Takes a character vector of 4 values to be plotted as a banner}
}
\value{
Saves a 'mydna.png' banner in the Desktop directory
}
\description{
User can pass in 4 single traits of a character type inside a "c()" function & a banner will be generated & saved in the Desktop.
}
